+++ src/components/content-display/social-media/flickr/flickr-oembed.tsx
@@ -1,6 +1,4 @@
-import Script from "next/script";
-import { UnknownObject } from "../../new-article/types";
-import { flickrConfig } from "./flickr.config";
+import { UnknownObject } from "@/src/types";
 
 interface OembedObject {
 	searchParams: UnknownObject;
