+++ src/components/content-display/social-media/mix-cloud/mix-cloud-oembed.tsx
@@ -1,5 +1,5 @@
 import Script from "next/script";
-import { UnknownObject } from "../../new-article/types";
+import { UnknownObject } from "@/src/types";
 import { mixCloudConfig } from "./mix-cloud.config";
 
 interface OembedObject {
