+++ src/components/content-display/social-media/rumble/rumble-oembed.tsx
@@ -1,5 +1,5 @@
 import Script from "next/script";
-import { UnknownObject } from "../../new-article/types";
+import { UnknownObject } from "@/src/types";
 import { rumbleConfig } from "./rumble.config";
 
 interface OembedObject {
