+++ src/components/content-display/social-media/smash-notes/smash-notes-oembed.tsx
@@ -1,5 +1,5 @@
 import Script from "next/script";
-import { UnknownObject } from "../../new-article/types";
+import { UnknownObject } from "@/src/types";
 import { smashNotesConfig } from "./smash-notes.config";
 
 interface OembedObject {
