+++ src/components/content-display/social-media/spotify/spotify-oembed.tsx
@@ -1,5 +1,5 @@
 import Script from "next/script";
-import { UnknownObject } from "../../new-article/types";
+import { UnknownObject } from "@/src/types";
 import { spotifyConfig } from "./spotify.config";
 
 interface OembedObject {
