+++ src/components/content-display/social-media/tiktok/tiktok-oembed.tsx
@@ -1,5 +1,5 @@
 import Script from "next/script";
-import { UnknownObject } from "../../new-article/types";
+import { UnknownObject } from "@/src/types";
 import { tikTokConfig } from "./tiktok.config";
 
 interface OembedObject {
