+++ src/components/content-display/social-media/vimeo/vimeo-oembed.tsx
@@ -1,5 +1,5 @@
 import Script from "next/script";
-import { UnknownObject } from "../../new-article/types";
+import { UnknownObject } from "@/src/types";
 import { vimeoConfig } from "./vimeo.config";
 
 interface OembedObject {
