+++ src/components/content-display/social-media/x/x-oembed.tsx
@@ -1,5 +1,5 @@
 import Script from "next/script";
-import { UnknownObject } from "../../new-article/types";
+import { UnknownObject } from "@/src/types";
 import { Xconfig } from "./x.config";
 
 interface OembedObject {
